services/task-service/pyproject.toml
[project]
name = "task-service"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "alembic==1.13.2",
    "greenlet==3.0.3",
    "dishka==1.4.2",
    "httpx==0.26.0",
    "asyncpg==0.30.0",
    "colorlog==6.8.2",
    "faststream[kafka]==0.5.12",
    "prometheus-fastapi-instrumentator==7.1.0",
    "pyjwt==2.8.0",
    "tenacity==9.1.4",
]

[project.optional-dependencies]
dev = [
    "flake8==6.1.0",
    "mypy==1.7.1",
    "black==23.11.0",
    "isort==5.12.0",
    "autopep8==2.0.4",
    "pytest-env==1.1.3",
    "pytest-cov==5.0.0",
    "pre-commit==3.8.0",
]
